Don't run server nor phantom if `test` is in args
We might want to spawn the server and phantomjs only once on task
startup, while executing the tests multiple times later.

A usecase is with a watch task monitoring file changes where we don't
want the overhead of starting the server and binding phantomjs each time
the tests are run.

Another usecase is binding multiple browsers manually to the server,
which is impossible if the server is relaunched for each test run.