Fix tests for Node v0.8.x
It seems like a timeout of `0` is a little too eager for Node v0.8.
Probably as the order of timeouts added to the event loop are executed
at random (which is sensible, but doesn't seem to be the case with
0.10.x). The deferred callbacks seem to be called after the test's
asynchronous assertion functions, causing them to break. Seem like
adding a few milliseconds of lag helps the event loop decide on what to
evaluate first.